This commit introduces the use of temporary objects to the canvas
Why temporary objects? As the name suggests, they are not meant to stay on the canvas for long and are used for quick interacting with the canvas. As a result we could easily add features like dynamically linking different non-temporary objects, and it became easy to add an icon related to the non-temporary object that allows you to dynamically remove objects from the canvas. But the changes were not only done on the front end, as the changes have to be carefully processed and managed in the background to update the connections information on all non-temporary objects and manage the display of temporary objects. Hope you enjoy these new features